backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
sqlalchemy==2.0.25
pytesseract==0.3.13
Pillow==10.4.0 